requirements.txt
# Local voice assistant (CLI) — no API keys required
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
pyttsx3>=2.90
openai-whisper>=20231117

# Web app (main.py): API + search + calendar
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
ddgs>=9.0.0
beautifulsoup4>=4.12.0
httpx>=0.26.0
anthropic>=0.39.0
google-api-python-client>=2.110.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

